Add function type QV for validation functions
Some of our quoting functions cannot guarantee lossless encoding of all
input due to limitations of the target parser, and so for these we'll
introduce sidecar validation functions to determine if a particular
command line is valid for the function in question.

These are given as separate functions rather than simply another return
value from Q because the translation functions should still be designed
so that they lose only meaningless information and so some applications
can reasonably ignore the fact that some input is invalid.