ViaPowerShell function
Since PowerShell itself is just an intermediary to launching a child
process, it's more like our handling of cmd.exe and VTs than it is a
quoting function. However, it requires us to isolate the program name from
the rest of the command line, so we need to introduce a new concept in
order to support that while retaining the ability to swap out different
wrapped quoting functions.